+++ routes/storage.php
@@ -17,7 +17,7 @@ $servicePattern = '[_0-9a-zA-Z-.]+';
 if (empty(config('df.storage_route_prefix'))) {
     // without a route prefix, this needs to use valid service names, otherwise prohibits any additional routes
     $service = Request::segment(1);
-    $storageServices = DreamFactory\Core\Facades\ServiceManager::getServiceNames(true, DreamFactory\Core\Enums\ServiceTypeGroups::FILE);
+    $storageServices = DreamFactory\Core\Facades\ServiceManager::getServiceNamesByGroup(DreamFactory\Core\Enums\ServiceTypeGroups::FILE, true);
     if (false !== array_search($service, $storageServices)) {
         Route::get('{storage}/{path}', 'StorageController@handleGET')->where(
             ['storage' => $servicePattern, 'path' => $resourcePathPattern]